+++ CHANGELOG.md
@@ -41,6 +41,7 @@ version of DSCResource.Common
 - Added CMClientSettingsStateMessaging resource
 - Added CMClientSettingsUserDeviceAffinity resource
 - Added CMSiteConfiguration resource
+- Added CMHierarchySetting resource
 
 ### Fixed